Extract desktop/src/projects.js (#17 Stage 4b)
Carve out the pure path computations that name-scope a project:

  projectRoot, projectAiDir, projectRunsDir, projectDesktopDir,
  projectIndexFile, terminalHistoryFile, aidevScript

projects.js exports a `createProjectPaths({ getSettings, appRoot })`
factory; main.js wires it to its own `settings()` and `APP_ROOT`,
destructures the bound functions, and keeps every call site working
unchanged. `projectConfig` stays in main.js because it still calls
`ensureWorkspace`, which depends on app data dirs not yet extracted.

Numbers:
- main.js: 1843 -> 1826 (-17, -62 cumulative since Stage 4a)
- projects.js: 45 lines (new)

Verified:
- npm run check: clean
- npm run smoke: 71 unit tests + 12 behavior smoke pass

Refs #17.
